commented out window scroll effect. due to an unintentional collapsing of navbar when on mobile view. see notes
In the production version of the portfolio, the navbar is toggled to collapse in random positions in the window. The intentional behaviour is to collapse the navbar on scroll but in the productio version the navbar is coillapsed immeditatly after toggle. Currently cannot find solution. Will continue to search for solution. For now navbar only collapses on click.